requirements.txt
absl-py==1.4.0
astunparse==1.6.3
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cvlib==0.2.7
Flask==2.3.2
Flask-MySQL==1.5.2
Flask-Session==0.5.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
idna==3.4
imageio==2.31.1
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.13.1
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.0
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
paho-mqtt==1.6.1
Pillow==10.0.0
progressbar==2.5
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyMySQL==1.1.0
pymysql-pool==0.4.3
PyMysqlPool==1.0.9
pyzbar==0.1.9
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.5.0
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
